Remove `pointOfSaleAsATabi2` feature flag and separate POS tab visibility and eligibility.
