+++ Modules/Sources/Experiments/DefaultFeatureFlagService.swift
@@ -88,8 +88,6 @@ public struct DefaultFeatureFlagService: FeatureFlagService {
             return true
         case .productImageOptimizedHandling:
             return true
-        case .pointOfSaleAsATabi2:
-            return true
         case .pointOfSaleOrdersi1:
             return true
         case .pointOfSaleOrdersi2:
