+++ WooCommerce/Classes/POS/Controllers/POSEntryPointController.swift
@@ -1,20 +1,19 @@
 import SwiftUI
-import protocol Experiments.FeatureFlagService
+
+protocol POSEntryPointEligibilityCheckerProtocol {
+    /// Determines whether the site is eligible for POS.
+    func checkEligibility() async -> POSEligibilityState
+    /// Refreshes the eligibility state based on the provided ineligible reason.
+    func refreshEligibility(ineligibleReason: POSIneligibleReason) async throws -> POSEligibilityState
+}
 
 @Observable final class POSEntryPointController {
     private(set) var eligibilityState: POSEligibilityState?
     private let posEligibilityChecker: POSEntryPointEligibilityCheckerProtocol
-    private let featureFlagService: POSFeatureFlagProviding
 
-    init(eligibilityChecker: POSEntryPointEligibilityCheckerProtocol,
-         featureFlagService: POSFeatureFlagProviding) {
+    init(eligibilityChecker: POSEntryPointEligibilityCheckerProtocol) {
         self.posEligibilityChecker = eligibilityChecker
-        self.featureFlagService = featureFlagService
 
-        guard featureFlagService.isFeatureFlagEnabled(.pointOfSaleAsATabi2) else {
-            self.eligibilityState = .eligible
-            return
-        }
         Task { @MainActor in
             eligibilityState = await posEligibilityChecker.checkEligibility()
         }
