+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/POS/POSTabVisibilityChecker.swift
@@ -1,0 +1,164 @@
+import Foundation
+import UIKit
+import class WooFoundation.CurrencySettings
+import enum WooFoundation.CountryCode
+import enum WooFoundation.CurrencyCode
+import protocol Experiments.FeatureFlagService
+import struct Yosemite.SiteSetting
+import struct Yosemite.Site
+import protocol Yosemite.POSEligibilityServiceProtocol
+import protocol Yosemite.StoresManager
+import class Yosemite.POSEligibilityService
+import enum Yosemite.FeatureFlagAction
+import class Yosemite.SiteAddress
+
+final class POSTabVisibilityChecker: POSTabVisibilityCheckerProtocol {
+    private let site: Site
+    private let userInterfaceIdiom: UIUserInterfaceIdiom
+    private let siteSettings: SelectedSiteSettingsProtocol
+    private let eligibilityService: POSEligibilityServiceProtocol
+    private let stores: StoresManager
+    private let featureFlagService: FeatureFlagService
+    private let siteCIABEligibilityChecker: CIABEligibilityCheckerProtocol
+
+    init(site: Site,
+         userInterfaceIdiom: UIUserInterfaceIdiom = UIDevice.current.userInterfaceIdiom,
+         siteSettings: SelectedSiteSettingsProtocol = ServiceLocator.selectedSiteSettings,
+         eligibilityService: POSEligibilityServiceProtocol = POSEligibilityService(),
+         stores: StoresManager = ServiceLocator.stores,
+         featureFlagService: FeatureFlagService = ServiceLocator.featureFlagService,
+         siteCIABEligibilityChecker: CIABEligibilityCheckerProtocol = CIABEligibilityChecker()) {
+        self.site = site
+        self.userInterfaceIdiom = userInterfaceIdiom
+        self.siteSettings = siteSettings
+        self.eligibilityService = eligibilityService
+        self.stores = stores
+        self.featureFlagService = featureFlagService
+        self.siteCIABEligibilityChecker = siteCIABEligibilityChecker
+    }
+
+    /// Checks the initial visibility of the POS tab without dependance on network requests.
+    func checkInitialVisibility() -> Bool {
+        eligibilityService.loadCachedPOSTabVisibility(siteID: site.siteID) ?? false
+    }
+
+    /// Checks the final visibility of the POS tab.
+    func checkVisibility() async -> Bool {
+        guard siteCIABEligibilityChecker.isFeatureSupported(.pointOfSale, for: site) else {
+            return false
+        }
+
+        guard userInterfaceIdiom == .pad else {
+            return false
+        }
+
+        async let siteSettingsEligibility = waitAndCheckSiteSettingsEligibility()
+        async let featureFlagEligibility = checkRemoteFeatureEligibility()
+
+        switch await siteSettingsEligibility {
+        case .ineligible(.unsupportedCountry):
+            return false
+        default:
+            break
+        }
+
+        return await featureFlagEligibility == .eligible
+    }
+}
+
+// MARK: - Site Settings Related Eligibility Check
+
+private extension POSTabVisibilityChecker {
+    enum SiteSettingsEligibilityState {
+        case eligible
+        case ineligible(reason: SiteSettingsIneligibleReason)
+    }
+
+    enum SiteSettingsIneligibleReason {
+        case siteSettingsNotAvailable
+        case unsupportedCountry(supportedCountries: [CountryCode])
+        case unsupportedCurrency(countryCode: CountryCode, supportedCurrencies: [CurrencyCode])
+    }
+
+    func waitAndCheckSiteSettingsEligibility() async -> SiteSettingsEligibilityState {
+        // Waits for the first site settings that matches the given site ID.
+        let siteSettings = await waitForSiteSettingsRefresh()
+        guard siteSettings.isNotEmpty else {
+            return .ineligible(reason: .siteSettingsNotAvailable)
+        }
+
+        // Conditions that can change if site settings are synced during the lifetime.
+        let countryCode = SiteAddress(siteSettings: siteSettings).countryCode
+        let currencyCode = CurrencySettings(siteSettings: siteSettings).currencyCode
+
+        return isEligibleFromCountryAndCurrencyCode(countryCode: countryCode, currencyCode: currencyCode)
+    }
+
+    func waitForSiteSettingsRefresh() async -> [SiteSetting] {
+        for await siteSettings in siteSettings.settingsStream.values {
+            guard siteSettings.siteID == site.siteID, siteSettings.settings.isNotEmpty, siteSettings.source != .initialLoad else {
+                continue
+            }
+            return siteSettings.settings
+        }
+        // If we get here, the stream completed without yielding any values for our site ID which is unexpected.
+        return []
+    }
+
+    func isEligibleFromCountryAndCurrencyCode(countryCode: CountryCode, currencyCode: CurrencyCode) -> SiteSettingsEligibilityState {
+        let supportedCountries: [CountryCode] = [.US, .GB]
+        let supportedCurrencies: [CountryCode: [CurrencyCode]] = [.US: [.USD],
+                                                                  .GB: [.GBP]]
+
+        // Checks country first.
+        guard supportedCountries.contains(countryCode) else {
+            return .ineligible(reason: .unsupportedCountry(supportedCountries: supportedCountries))
+        }
+
+        let supportedCurrenciesForCountry = supportedCurrencies[countryCode] ?? []
+        guard supportedCurrenciesForCountry.contains(currencyCode) else {
+            return .ineligible(reason: .unsupportedCurrency(countryCode: countryCode, supportedCurrencies: supportedCurrenciesForCountry))
+        }
+        return .eligible
+    }
+}
+
+// MARK: - Remote Feature Flag Eligibility Check
+
+private extension POSTabVisibilityChecker {
+    enum RemoteFeatureFlagEligibilityState: Equatable {
+        case eligible
+        case ineligible(reason: RemoteFeatureFlagIneligibleReason)
+    }
+
+    enum RemoteFeatureFlagIneligibleReason: Equatable {
+        case selfDeallocated
+        case featureFlagDisabled
+    }
+
+    @MainActor
+    func checkRemoteFeatureEligibility() async -> RemoteFeatureFlagEligibilityState {
+        // Only whitelisted accounts in WPCOM have the Point of Sale remote feature flag enabled. These can be found at D159901-code
+        // If the account is whitelisted, then the remote value takes preference over the local feature flag configuration
+        await withCheckedContinuation { [weak self] continuation in
+            guard let self else {
+                return continuation.resume(returning: .ineligible(reason: .selfDeallocated))
+            }
+            let action = FeatureFlagAction.isRemoteFeatureFlagEnabled(.pointOfSale, defaultValue: false) { [weak self] result in
+                guard let self else {
+                    return continuation.resume(returning: .ineligible(reason: .selfDeallocated))
+                }
+                switch result {
+                case true:
+                    // The site is whitelisted.
+                    continuation.resume(returning: .eligible)
+                case false:
+                    // When the site is not whitelisted, check the local feature flag configuration.
+                    let localFeatureFlag = featureFlagService.isFeatureFlagEnabled(.pointOfSale)
+                    continuation.resume(returning: localFeatureFlag ? .eligible : .ineligible(reason: .featureFlagDisabled))
+                }
+            }
+            self.stores.dispatch(action)
+        }
+    }
+}
