+++ WooCommerce/WooCommerceTests/Mocks/MockFeatureFlagService.swift
@@ -14,7 +14,6 @@ final class MockFeatureFlagService: FeatureFlagService, POSFeatureFlagProviding 
     var isSubscriptionsInOrderCreationCustomersEnabled: Bool
     var isSubscriptionsInOrderCreationUIEnabled: Bool
     var isPointOfSaleEnabled: Bool
-    var isPointOfSaleAsATabi2Enabled: Bool
     var googleAdsCampaignCreationOnWebView: Bool
     var blazeEvergreenCampaigns: Bool
     var blazeCampaignObjective: Bool
@@ -39,7 +38,6 @@ final class MockFeatureFlagService: FeatureFlagService, POSFeatureFlagProviding 
          isSubscriptionsInOrderCreationCustomersEnabled: Bool = false,
          isSubscriptionsInOrderCreationUIEnabled: Bool = false,
          isPointOfSaleEnabled: Bool = false,
-         isPointOfSaleAsATabi2Enabled: Bool = false,
          googleAdsCampaignCreationOnWebView: Bool = false,
          blazeEvergreenCampaigns: Bool = false,
          blazeCampaignObjective: Bool = false,
@@ -62,7 +60,6 @@ final class MockFeatureFlagService: FeatureFlagService, POSFeatureFlagProviding 
         self.isSubscriptionsInOrderCreationCustomersEnabled = isSubscriptionsInOrderCreationCustomersEnabled
         self.isSubscriptionsInOrderCreationUIEnabled = isSubscriptionsInOrderCreationUIEnabled
         self.isPointOfSaleEnabled = isPointOfSaleEnabled
-        self.isPointOfSaleAsATabi2Enabled = isPointOfSaleAsATabi2Enabled
         self.googleAdsCampaignCreationOnWebView = googleAdsCampaignCreationOnWebView
         self.blazeEvergreenCampaigns = blazeEvergreenCampaigns
         self.blazeCampaignObjective = blazeCampaignObjective
@@ -107,8 +104,6 @@ final class MockFeatureFlagService: FeatureFlagService, POSFeatureFlagProviding 
             return isSubscriptionsInOrderCreationUIEnabled
         case .pointOfSale:
             return isPointOfSaleEnabled
-        case .pointOfSaleAsATabi2:
-            return isPointOfSaleAsATabi2Enabled
         case .googleAdsCampaignCreationOnWebView:
             return googleAdsCampaignCreationOnWebView
         case .blazeEvergreenCampaigns:
