+++ WooCommerce/WooCommerceTests/Mocks/MockPOSTabVisibilityChecker.swift
@@ -1,0 +1,16 @@
+import Foundation
+@testable import WooCommerce
+
+final class MockPOSTabVisibilityChecker: POSTabVisibilityCheckerProtocol {
+    var initialVisibility: Bool = false
+    var visibility: Bool = false
+
+    func checkInitialVisibility() -> Bool {
+        initialVisibility
+    }
+
+    @MainActor
+    func checkVisibility() async -> Bool {
+        visibility
+    }
+}
